backend/requirements.txt
Django>=4.2
mysqlclient>=2.0
djangorestframework
django-cors-headers
